feat: publish recipe_upserted and recipe_deleted

+++ services/recipes/lib/src/controllers/recipe.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:mongo_dart/mongo_dart.dart';
 import 'package:recipes/recipes.dart';
 import 'package:shelf_plus/shelf_plus.dart';
 
@@ -24,24 +25,26 @@ class RecipeController {
     Request request,
   ) async {
     //
-    try {
-      final json = await request.body.asJson;
+    final json = await request.body.asJson;
 
-      final form = RecipeForm.fromJson(json);
+    final form = RecipeForm.fromJson(json);
 
-      final dao = RecipeDao();
-      final data = await dao.post(form);
+    final dao = RecipeDao();
+    final data = await dao.post(form);
 
-      return Response(201, body: jsonEncode(data));
-    } //
-    on TypeError catch (e) {
-      final error = NlIruoyCommonModelsError(
-        code: 'invalid-json',
-        message: e.toString(),
-      );
+    // Publish the upserted event for other services.
+    final event = {
+      'recipe_upserted': RecipeUpserted(
+        id: ObjectId().$oid,
+        timestamp: DateTime.now(),
+        recipe: data,
+      ).toJson(),
+    };
 
-      return Response(400, body: jsonEncode(error));
-    }
+    final queue = await getQueue();
+    await queue.add(RecipeEvent.fromJson(event));
+
+    return Response(201, body: jsonEncode(data));
   }
 
   static Future<Response> putById(
@@ -49,24 +52,33 @@ class RecipeController {
     String id,
   ) async {
     //
-    try {
-      final json = await request.body.asJson;
+    final json = await request.body.asJson;
+    final form = RecipeForm.fromJson(json);
 
-      final form = RecipeForm.fromJson(json);
+    final dao = RecipeDao();
 
-      final dao = RecipeDao();
-      final data = await dao.putById(form, id: id);
+    final existing = await dao.getById(id: id);
 
-      return Response(200, body: jsonEncode(data));
-    } //
-    on TypeError catch (e) {
-      final error = NlIruoyCommonModelsError(
-        code: 'invalid-json',
-        message: e.toString(),
-      );
-
-      return Response(400, body: jsonEncode(error));
+    if (existing == null) {
+      // The recipe was not found and can therefor not be deleted.
+      return Response(404);
     }
+
+    final data = await dao.putById(form, id: id);
+
+    // Publish the upserted event for other services.
+    final event = {
+      'recipe_upserted': RecipeUpserted(
+        id: ObjectId().$oid,
+        timestamp: DateTime.now(),
+        recipe: data,
+      ).toJson(),
+    };
+
+    final queue = await getQueue();
+    await queue.add(RecipeEvent.fromJson(event));
+
+    return Response(200, body: jsonEncode(data));
   }
 
   static Future<Response> deleteById(
@@ -75,7 +87,27 @@ class RecipeController {
   ) async {
     //
     final dao = RecipeDao();
+
+    final existing = await dao.getById(id: id);
+
+    if (existing == null) {
+      // The recipe was not found and can therefor not be deleted.
+      return Response(404);
+    }
+
     await dao.deleteById(id: id);
+
+    // Publish the deleted event for other services.
+    final event = {
+      'recipe_deleted': RecipeDeleted(
+        id: ObjectId().$oid,
+        timestamp: DateTime.now(),
+        recipe: existing,
+      ).toJson(),
+    };
+
+    final queue = await getQueue();
+    await queue.add(RecipeEvent.fromJson(event));
 
     return Response(204);
   }
